+++ src/date/timezone.c
@@ -1,4 +1,4 @@
 #include "timezone.h"
 
 
-int16 implicit_timezone = 0;
+int32 implicit_timezone = 0;
